Add useHintLayout hook to manage hint layout state and measurements
